packages/bots/axiom-assist-bot/secret-agent/requirements.txt
# secret-agent/requirements.txt - Dependencies for Secret Agent Cloud Function
google-cloud-secret-manager>=2.16.0
requests>=2.31.0